feat(repository): automate update logging in AuctionHouseDynamoRepository
Integrate AuctionHouseUpdateLogDynamoRepository to automatically record log entries whenever an auction house is saved. This change also updates the size field from Int to Double for better precision and adds validation for mandatory fields in the repository layer.

+++ src/main/kotlin/net/jonasmf/auctionengine/domain/AuctionHouseUpdateLog.kt
@@ -5,7 +5,7 @@ import kotlin.time.Instant
 class AuctionHouseUpdateLog(
     val id: Int,
     val lastModified: Instant,
-    val size: Int,
+    val size: Double,
     val timeSincePreviousDump: Long,
     val url: String,
 )
